feat(stage 2): Pattern B migrate branch_eq + branch_lt + bigint Branch256
Switch the branch family from u8-shaped (Pattern A) to u16-shaped
(Pattern B): each chip carries fewer columns and range-checks each limb
once via the variable range checker bus instead of byte-pair-checking
via the bitwise lookup bus.

Cores:
- `BranchEqualCoreAir` and `BranchEqualCoreRecord` now parameterize on
  `LIMB_BITS`; records are `[u16; NUM_LIMBS]`; AIR range-checks each
  read limb to `[0, 2^LIMB_BITS)` via `VariableRangeCheckerBus`.
- `BranchLessThanCoreAir` same treatment. MSB sign-shift, non-MSB
  pair checks, and the diff-nonzero check all use per-limb
  `range_check`.

Adapters:
- `Rv64BranchAdapter` reads `[[u16; BLOCK_FE_WIDTH]; 2]` natively via a
  new `tracing_read_u16` helper (byte-pointer in, cell-indexed
  `memory.read::<u16, N>` under the hood). Bridge call passes the
  u16 expressions through directly, no `pack_u8_for_bus`.
- `Rv64VecHeapBranchAdapter` `READ_SIZE` is now in u16 cells. Each heap
  block is `BLOCK_FE_WIDTH=4` u16 cells (= 8 bytes); the AIR's
  per-block byte stride is `BUS_PTR_SCALE * READ_SIZE = 8`.
- `VecToFlatBranchAdapterExecutor` flattens `[[[u16; BLOCK_SIZE]; BPR];
  NR]` → `[[u16; TOTAL_READ_SIZE]; NR]`.

Type aliases:
- RV64: `RV64_BRANCH_NUM_LIMBS = BLOCK_FE_WIDTH = 4`,
  `RV64_BRANCH_LIMB_BITS = 16`.
- bigint: `INT256_NUM_U16_LIMBS = 16`, `INT256_U16_LIMB_BITS = 16`;
  branch chips drop INT256_NUM_LIMBS/RV64_CELL_BITS in favor of these.

Extension wiring (RV64 + bigint, both CPU and CUDA): branch chips now
receive the `VariableRangeChecker` shared chip; they no longer need
the bitwise op lookup chip.

CUDA: `beq.cuh`/`blt.cuh` cores parameterize on `LIMB_BITS` and use
`VariableRangeChecker.add_count` per limb. `beq.cu`/`blt.cu` and the
bigint `bigint.cu` instantiate the new shapes and drop the
bitwise-lookup parameter from the kernel and ABI signatures.

Tests: register data generated as `[u16; 4]` (RV64) / `[u8; 32]`
(bigint, byte-view); the test writes via `tester.write::<8>(...)` so
the u8 byte-view path serializes the values into u16-celled memory
backing. Negative tests pranked accordingly.

CPU verified: 318/318 openvm-riscv-circuit tests pass, 6/6
openvm-bigint-circuit branch tests pass.

+++ extensions/riscv-adapters/src/vec_to_flat.rs
@@ -487,12 +487,12 @@ where
     F: PrimeField32,
     A: AdapterTraceExecutor<
         F,
-        ReadData = [[[u8; BLOCK_SIZE]; BLOCKS_PER_READ]; NUM_READS],
+        ReadData = [[[u16; BLOCK_SIZE]; BLOCKS_PER_READ]; NUM_READS],
         WriteData = (),
     >,
 {
     const WIDTH: usize = A::WIDTH;
-    type ReadData = [[u8; TOTAL_READ_SIZE]; NUM_READS];
+    type ReadData = [[u16; TOTAL_READ_SIZE]; NUM_READS];
     type WriteData = ();
     type RecordMut<'a>
         = A::RecordMut<'a>
@@ -514,7 +514,7 @@ where
         let data_inner = <A as AdapterTraceExecutor<F>>::read(&self.0, memory, instruction, record);
 
         core::array::from_fn(|i| {
-            let mut out = [0u8; TOTAL_READ_SIZE];
+            let mut out = [0u16; TOTAL_READ_SIZE];
             for (block_idx, block) in data_inner[i].iter().enumerate() {
                 let start = block_idx * BLOCK_SIZE;
                 out[start..start + BLOCK_SIZE].copy_from_slice(&block[..]);
